Initial spectrogram implementation + refactor
Reorder things to make the code tidier and add a spectrogram
implementation. This is essentially taking the stft magnitude spectrum
and applying a power to it. Also made the FrequencyComponent trait more
generic
