+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Short time fourier transform
-- Preemphasis function
+- Spectogram extension trait
+- Preemphasis extension trait
 - [Internal] Criterion for benchmarking and a benchmark for the STFT
 - [Internal] cargo-fuzz test for STFT that takes a corpus of wav files
